+++ tests/PhpSpreadsheetTests/Calculation/Functions/Statistical/PermutTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\Statistical;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\Statistical;
+use PHPUnit\Framework\TestCase;
+
+class PermutTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerPERMUT
+     *
+     * @param mixed $expectedResult
+     */
+    public function testPERMUT($expectedResult, ...$args)
+    {
+        $result = Statistical::PERMUT(...$args);
+        self::assertEquals($expectedResult, $result, '', 1E-12);
+    }
+
+    public function providerPERMUT()
+    {
+        return require 'data/Calculation/Statistical/PERMUT.php';
+    }
+}
