+++ tests/PhpSpreadsheetTests/Calculation/StatisticalTest.php
@@ -190,122 +190,6 @@ class StatisticalTest extends TestCase
     }
 
     /**
-     * @dataProvider providerBasicCOUNT
-     *
-     * @param mixed $expectedResult
-     */
-    public function testBasicCOUNT($expectedResult, ...$args)
-    {
-        $result = Statistical::COUNT(...$args);
-        self::assertEquals($expectedResult, $result, '', 1E-12);
-    }
-
-    public function providerBasicCOUNT()
-    {
-        return require 'data/Calculation/Statistical/BasicCOUNT.php';
-    }
-
-    /**
-     * @dataProvider providerExcelCOUNT
-     *
-     * @param mixed $expectedResult
-     */
-    public function testExcelCOUNT($expectedResult, ...$args)
-    {
-        $result = Statistical::COUNT(...$args);
-        self::assertEquals($expectedResult, $result, '', 1E-12);
-    }
-
-    public function providerExcelCOUNT()
-    {
-        return require 'data/Calculation/Statistical/ExcelCOUNT.php';
-    }
-
-    /**
-     * @dataProvider providerOpenOfficeCOUNT
-     *
-     * @param mixed $expectedResult
-     */
-    public function testOpenOfficeCOUNT($expectedResult, ...$args)
-    {
-        Functions::setCompatibilityMode(Functions::COMPATIBILITY_OPENOFFICE);
-
-        $result = Statistical::COUNT(...$args);
-        self::assertEquals($expectedResult, $result, '', 1E-12);
-    }
-
-    public function providerOpenOfficeCOUNT()
-    {
-        return require 'data/Calculation/Statistical/OpenOfficeCOUNT.php';
-    }
-
-    /**
-     * @dataProvider providerGnumericCOUNT
-     *
-     * @param mixed $expectedResult
-     */
-    public function testGnumericCOUNT($expectedResult, ...$args)
-    {
-        Functions::setCompatibilityMode(Functions::COMPATIBILITY_GNUMERIC);
-
-        $result = Statistical::COUNT(...$args);
-        self::assertEquals($expectedResult, $result, '', 1E-12);
-    }
-
-    public function providerGnumericCOUNT()
-    {
-        return require 'data/Calculation/Statistical/GnumericCOUNT.php';
-    }
-
-    /**
-     * @dataProvider providerCOUNTA
-     *
-     * @param mixed $expectedResult
-     */
-    public function testCOUNTA($expectedResult, ...$args)
-    {
-        $result = Statistical::COUNTA(...$args);
-        self::assertEquals($expectedResult, $result, '', 1E-12);
-    }
-
-    public function providerCOUNTA()
-    {
-        return require 'data/Calculation/Statistical/COUNTA.php';
-    }
-
-    /**
-     * @dataProvider providerCOUNTIF
-     *
-     * @param mixed $expectedResult
-     */
-    public function testCOUNTIF($expectedResult, ...$args)
-    {
-        $result = Statistical::COUNTIF(...$args);
-        self::assertEquals($expectedResult, $result, '', 1E-12);
-    }
-
-    public function providerCOUNTIF()
-    {
-        return require 'data/Calculation/Statistical/COUNTIF.php';
-    }
-
-    /**
-     * @dataProvider providerCOUNTIFS
-     *
-     * @param mixed $expectedResult
-     */
-    public function testCOUNTIFS($expectedResult, ...$args)
-    {
-        $result = Statistical::COUNTIFS(...$args);
-        self::assertEquals($expectedResult, $result, '', 1E-12);
-    }
-
-    public function providerCOUNTIFS()
-    {
-        return require 'data/Calculation/Statistical/COUNTIFS.php';
-    }
-
-    /**
      * @dataProvider providerCOVAR
      *
      * @param mixed $expectedResult
