+++ tests/data/Calculation/Statistical/COUNTBLANK.php
@@ -1,0 +1,8 @@
+<?php
+
+return [
+    [
+        4,
+        [-1, null, 0, '', 1, null, 2, '', 3],
+    ],
+];
