+++ tests/data/Calculation/Statistical/PERMUT.php
@@ -1,0 +1,48 @@
+<?php
+
+return [
+    [
+        5,
+        5, 1,
+    ],
+    [
+        20,
+        5, 2,
+    ],
+    [
+        60,
+        5, 3,
+    ],
+    [
+        120,
+        5, 4,
+    ],
+    [
+        120,
+        5, 5,
+    ],
+    [
+        970200,
+        100, 3,
+    ],
+    [
+        6,
+        3, 2,
+    ],
+    [
+        720,
+        6, 6,
+    ],
+    [
+        5040,
+        7, 6,
+    ],
+    [
+        151200,
+        10, 6,
+    ],
+    [
+        10068347520,
+        49,6,
+    ],
+];
